+++ logprep/util/prometheus_exporter.py
@@ -5,22 +5,21 @@ import tempfile
 from os import listdir, path
 from os.path import isfile
 
-from prometheus_client import start_http_server, multiprocess, REGISTRY, Gauge
+from prometheus_client import REGISTRY, Gauge, multiprocess, start_http_server
 
 
 class PrometheusStatsExporter:
     """Used to control the prometheus exporter and to manage the metrics"""
 
     metric_prefix: str = "logprep_"
-    multi_processing_dir = None
+    multi_processing_dir: str = None
 
     def __init__(self, status_logger_config, application_logger):
         self._logger = application_logger
         self.configuration = status_logger_config
-        self._port = 8000
+        self._port = status_logger_config.get("port", 8000)
 
         self._prepare_multiprocessing()
-        self._extract_port_from(self.configuration)
         self._set_up_metrics()
 
     def _prepare_multiprocessing(self):
@@ -60,19 +59,13 @@ class PrometheusStatsExporter:
                 removed_files.append(filename)
         self._logger.debug(f"Removed stale metric files: {removed_files}")
 
-    def _extract_port_from(self, configuration):
-        target_configs = configuration.get("targets", [])
-        for config in target_configs:
-            if "prometheus" in config:
-                self._port = config.get("prometheus").get("port")
-
     def _set_up_metrics(self):
         """Sets up the metrics that the prometheus exporter should expose"""
         self.metrics = {}
         self.tracking_interval = Gauge(
             f"{self.metric_prefix}tracking_interval_in_seconds",
             "Tracking interval",
-            labelnames=["component", "logprep_version", "config_version"],
+            labelnames=["component"],
             registry=None,
         )