+++ tests/unit/util/test_prometheus_exporter.py
@@ -8,7 +8,7 @@ from logging import getLogger
 from unittest import mock
 
 import pytest
-from prometheus_client import Gauge, REGISTRY
+from prometheus_client import REGISTRY, Gauge
 
 from logprep.util.prometheus_exporter import PrometheusStatsExporter
 
@@ -17,27 +17,24 @@ class TestPrometheusStatsExporter:
     def setup_method(self):
         REGISTRY.__init__()
         self.metrics_config = {
-            "period": 10,
-            "enabled": True,
-            "cumulative": True,
-            "targets": [
-                {"prometheus": {"port": 80}},
-                {"file": {"path": "", "rollover_interval": 200, "backup_count": 10}},
-            ],
+            "metrics": {"period": 10, "enabled": True, "cumulative": True, "port": 80}
         }
 
     def test_correct_setup(self):
-        exporter = PrometheusStatsExporter(self.metrics_config, getLogger("test-logger"))
+        exporter = PrometheusStatsExporter(
+            self.metrics_config.get("metrics"), getLogger("test-logger")
+        )
         assert not exporter.metrics
         assert isinstance(exporter.tracking_interval, Gauge)
-        assert exporter._port == self.metrics_config["targets"][0]["prometheus"]["port"]
+        assert exporter._port == self.metrics_config["metrics"]["port"]
 
     def test_default_port_if_missing_in_config(self):
         metrics_config = {
-            "period": 10,
-            "enabled": True,
-            "cumulative": True,
-            "targets": [{"file": {"path": "", "rollover_interval": 200, "backup_count": 10}}],
+            "metrics": {
+                "period": 10,
+                "enabled": True,
+                "cumulative": True,
+            }
         }
         exporter = PrometheusStatsExporter(metrics_config, getLogger("test-logger"))